fix(components/molecule/field): small error
